OcrManager, ImageProcessor 리팩토링 및 Type 정리
