--- src/main/modules/auth/auth.module.ts
@@ -1,12 +1,0 @@
-import { Module } from '@nestjs/common'
-import { FileManagerModule } from '../file-manager/file-manager.module'
-import { AuthController } from './auth.controller'
-import { AuthService } from './auth.service'
-
-@Module({
-  imports: [FileManagerModule],
-  controllers: [AuthController],
-  providers: [AuthService],
-  exports: [AuthService],
-})
-export class AuthModule {}
