+++ src/main/modules/file-manager/file-manager.service.ts
@@ -1,8 +1,8 @@
 import { Injectable, Logger } from '@nestjs/common'
-import type { SessionData } from '@src/types/common/SessionData'
-import type { SettingsData } from '@src/types/common/SettingData'
-import type { SongData } from '@src/types/common/SongData'
-import type { StorageInfo } from '@src/types/common/StroageInfo'
+import type { SongData } from '@src/types/games/SongData'
+import type { SessionData } from '@src/types/sessions/SessionData'
+import type { SettingsData } from '@src/types/settings/SettingData'
+import type { StorageInfo } from '@src/types/storage/StroageInfo'
 import { app, shell } from 'electron'
 import * as fs from 'fs-extra'
 import * as path from 'path'
