+++ src/main/modules/logger/logger.controller.ts
@@ -1,14 +1,14 @@
 import { IpcHandle } from '@doubleshot/nest-electron'
 import { Controller } from '@nestjs/common'
-import { LogLevel } from '@src/types/common/LogLevel'
+import { LogLevel } from '@src/types/dto/log/LogLevel'
 import { LoggerService } from './logger.service'
 
 @Controller()
 export class LoggerController {
-  constructor(private readonly loggingService: LoggerService) {}
+  constructor(private readonly loggerService: LoggerService) {}
 
   @IpcHandle('logger:create-log')
   public handlecreateLog(args: [LogLevel?, string?, ...unknown[]]): boolean {
-    return this.loggingService.createLog(...args)
+    return this.loggerService.createLog(...args)
   }
 }
