+++ src/main/modules/process-manager/process-manager.controller.ts
@@ -3,14 +3,14 @@ import { IpcHandle } from '@doubleshot/nest-electron'
 import { Controller } from '@nestjs/common'
 import { ProcessDescriptor } from 'ps-list'
 import { from, Observable } from 'rxjs'
-import { ProcessService } from './process.service'
+import { ProcessManagerService } from './process-manager.service'
 
 @Controller()
-export class ProcessController {
-  constructor(private readonly processService: ProcessService) {}
+export class ProcessManagerController {
+  constructor(private readonly processManagerService: ProcessManagerService) {}
 
   @IpcHandle('process:list')
   getPsList(): Observable<ProcessDescriptor[]> {
-    return from(this.processService.getPsList())
+    return from(this.processManagerService.getPsList())
   }
 }
