+++ src/main/modules/process-manager/process-manager.service.ts
@@ -3,7 +3,7 @@ import { Injectable } from '@nestjs/common'
 import { ProcessDescriptor } from 'ps-list'
 
 @Injectable()
-export class ProcessService {
+export class ProcessManagerService {
   public async getPsList(): Promise<ProcessDescriptor[]> {
     const psList = await import('ps-list')
     const processes = await psList.default()
