+++ src/preload/index.d.ts
@@ -1,7 +1,7 @@
-import type { LogLevel } from '@src/types/common/LogLevel'
-import type { SessionData } from '@src/types/common/SessionData'
-import type { SettingsData } from '@src/types/common/SettingData'
-import type { StorageInfo } from '@src/types/common/StroageInfo'
+import type { LogLevel } from '@src/types/dto/log/LogLevel'
+import type { SessionData } from '@src/types/sessions/SessionData'
+import type { SettingsData } from '@src/types/settings/SettingData'
+import type { StorageInfo } from '@src/types/storage/StroageInfo'
 import type { Result } from 'get-windows'
 import type { ProcessDescriptor } from 'ps-list'
