+++ src/render/components/ui/MenuSidebar.tsx
@@ -4,7 +4,7 @@ import { useAuth } from '@render/hooks/useAuth'
 import { useNotificationSystem } from '@render/hooks/useNotifications'
 import { RootState } from '@render/store'
 import { setIsOpenExternalLink, setOpenExternalLink } from '@render/store/slices/uiSlice'
-import type { MenuItem } from '@src/types/render/MenuItem'
+import type { MenuItem } from '@src/types/menu/MenuItem'
 import { motion } from 'framer-motion'
 import React, { useState } from 'react'
 import { useTranslation } from 'react-i18next'
