+++ src/render/components/ui/Notification.tsx
@@ -1,7 +1,7 @@
 import { Icon } from '@iconify/react'
 import { RootState } from '@render/store'
-import type { NotificationContainerProps } from '@src/types/render/NotificationContainerProps'
-import type { NotificationProps } from '@src/types/render/NotificationProps'
+import type { NotificationContainerProps } from '@src/types/notifications/NotificationContainerProps'
+import type { NotificationProps } from '@src/types/notifications/NotificationProps'
 import { AnimatePresence, motion } from 'framer-motion'
 import { useEffect, useState } from 'react'
 import { useTranslation } from 'react-i18next'
