+++ src/render/constants/buildInfo.ts
@@ -1,1 +1,1 @@
-export const BUILD_DATE = '202504281252 Client (Vite Developer Build)';
+export const BUILD_DATE = '202504281255 Client (Vite Developer Build)';
