+++ src/types/dto/ping/ServerStatusResponse.ts
@@ -1,4 +1,4 @@
-export interface ServerStatus {
+export interface ServerStatusResponse {
   message: string
   version: string
   timestamp: number
