+++ src/types/games/GameType.ts
