+++ src/types/notifications/NotificationProps.ts
@@ -1,0 +1,8 @@
+import type { Notification } from '@src/types/notifications/Notification'
+import type { NotificationContainerProps } from '@src/types/notifications/NotificationContainerProps'
+
+export interface NotificationProps {
+  notification: Notification
+  onRemove: NotificationContainerProps['onRemove']
+  index: number
+}
