+++ src/types/ocr/GameRegions.ts
@@ -1,0 +1,8 @@
+import { ProfileRegion } from './ProfileRegion'
+
+// 게임별 화면 타입에 따른 영역 인터페이스
+export interface GameRegions {
+  [key: string]: {
+    [screenType: string]: ProfileRegion
+  }
+}
