+++ src/types/ocr/OcrRegion.ts
@@ -1,0 +1,7 @@
+// OCR 영역 인터페이스
+export interface OCRRegion {
+  width: number
+  height: number
+  left: number
+  top: number
+}
