+++ src/types/ocr/OcrResultInfo.ts
@@ -1,0 +1,6 @@
+// OCR 결과 인터페이스
+export interface OCRResultInfo {
+  isResult: string[]
+  text: string
+  where: string
+}
