+++ src/types/ocr/ProfileRegion.ts
@@ -1,0 +1,8 @@
+import { OCRRegion } from './OcrRegion'
+
+// 프로파일 영역 인터페이스
+export interface ProfileRegion {
+  myProfile: OCRRegion
+  otherProfile: OCRRegion
+  chat: OCRRegion
+}
