+++ src/types/storage/StroageInfo.ts
