+++ src/types/tooltip/TooltipProps.ts
